+++ Changes.md
@@ -1,3 +1,4 @@
+* #1086 - Add playoff match number to team selection
 * #1088 - Assign head to head table labels before sending update to display
 * #1084 - Properly display head to head winners in the awards script
 * #1089 - Allow everyone to see the list of finalist teams